python/requirements.txt
click==6.6
Flask==0.11.1
Flask-GraphQL==1.3.1
graphene==1.1.3
graphql-core==1.0.1
graphql-relay==0.4.5
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
promise==1.0.1
six==1.10.0
typing==3.5.2.2
Werkzeug==0.11.11
